+++ cqproto/provider.go
@@ -3,6 +3,7 @@ package cqproto
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/cloudquery/cq-provider-sdk/provider/schema"
 	"github.com/cloudquery/cq-provider-sdk/provider/schema/diag"
@@ -52,8 +53,8 @@ type GetProviderSchemaResponse struct {
 	Version string
 	// ResourceTables is a map of tables this provider creates
 	ResourceTables map[string]*schema.Table
-	// Migrations scripts available for the provider
-	Migrations map[string][]byte
+	// Migrations scripts available for the provider, for all dialects
+	Migrations map[string]map[string][]byte
 }
 
 // GetProviderConfigRequest represents a CloudQuery RPC request for provider's config
@@ -70,8 +71,6 @@ type ConfigureProviderRequest struct {
 	Connection ConnectionDetails
 	// Config is the configuration the user supplied for the provider
 	Config []byte
-	// DisableDelete configures providers to skip deletion of data before resource fetch
-	DisableDelete bool
 	// Fields to inject to every resource on insert
 	ExtraFields map[string]interface{}
 }
@@ -182,3 +181,9 @@ func (p ProviderDiagnostic) Description() diag.Description {
 		Detail:   p.Details,
 	}
 }
+
+func (p ProviderDiagnostic) Error() string {
+	return fmt.Sprintf("%s: %s", p.ResourceName, p.Summary)
+}
+
+var _ diag.Diagnostic = (*ProviderDiagnostic)(nil)