+++ provider/schema/execution_test.go
@@ -24,7 +24,7 @@ var testMultiplexTable = &Table{
 	Multiplex: func(meta ClientMeta) []ClientMeta {
 		return []ClientMeta{meta}
 	},
-	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 		return nil
 	},
 	Columns: []Column{
@@ -39,7 +39,7 @@ var testMultiplexTable = &Table{
 			Multiplex: func(meta ClientMeta) []ClientMeta {
 				return []ClientMeta{meta}
 			},
-			Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+			Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 				return nil
 			},
 			Columns: []Column{
@@ -113,7 +113,7 @@ var testBadColumnResolverTable = &Table{
 			},
 		},
 	},
-	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 		res <- testDefaultsTableData{Name: nil}
 		return nil
 	},
@@ -145,35 +145,35 @@ var testIgnoreErrorColumnResolverTable = &Table{
 			},
 		},
 	},
-	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+	Resolver: func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 		res <- testDefaultsTableData{Name: nil}
 		return nil
 	},
 }
 
-func failingTableResolver(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+func failingTableResolver(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 	return fmt.Errorf("table resolve failed")
 }
 
-func doNothingResolver(_ context.Context, _ ClientMeta, _ *Resource, _ chan interface{}) error {
+func doNothingResolver(_ context.Context, _ ClientMeta, _ *Resource, _ chan<- interface{}) error {
 	return nil
 }
 
-func dataReturningResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan interface{}) error {
+func dataReturningResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan<- interface{}) error {
 	object := testTableStruct{}
 	_ = defaults.Set(&object)
 	res <- []testTableStruct{object, object, object}
 	return nil
 }
 
-func dataReturningSingleResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan interface{}) error {
+func dataReturningSingleResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan<- interface{}) error {
 	object := testTableStruct{}
 	_ = defaults.Set(&object)
 	res <- object
 	return nil
 }
 
-func passingNilResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan interface{}) error {
+func passingNilResolver(_ context.Context, _ ClientMeta, _ *Resource, res chan<- interface{}) error {
 	res <- nil
 	return nil
 }
@@ -191,22 +191,26 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 	t.Run("failed table resolver", func(t *testing.T) {
 		testTable.Resolver = failingTableResolver
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Error(t, err)
 	})
 
 	t.Run("failing table column resolver", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execFailing := NewExecutionData(mockDb, logger, testBadColumnResolverTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execFailing := NewExecutionData(mockDb, logger, testBadColumnResolverTable, nil, false)
 		_, err := execFailing.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Error(t, err)
 	})
 
 	t.Run("ignore error table column resolver w/partialFetch", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		exec := NewExecutionData(mockDb, logger, testIgnoreErrorColumnResolverTable, false, nil, true)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testIgnoreErrorColumnResolverTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testIgnoreErrorColumnResolverTable, exec.executionStart, mock.Anything).Return(nil)
 		var expectedResource *Resource
 		testIgnoreErrorColumnResolverTable.PostResourceResolver = func(ctx context.Context, meta ClientMeta, parent *Resource) error {
 			expectedResource = parent
@@ -220,14 +224,16 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 	})
 
 	t.Run("error table column resolver w/partialFetch", func(t *testing.T) {
-		testBadColumnResolverTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		testBadColumnResolverTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			someString := "noError"
 			res <- []testDefaultsTableData{{Name: &someString}, {Name: nil}, {Name: &someString}}
 			return nil
 		}
 		mockDb := new(DatabaseMock)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		exec := NewExecutionData(mockDb, logger, testBadColumnResolverTable, false, nil, true)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testBadColumnResolverTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testBadColumnResolverTable, exec.executionStart, mock.Anything).Return(nil)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Nil(t, err)
 		assert.Len(t, exec.PartialFetchFailureResult, 1)
@@ -235,7 +241,9 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("doing nothing resolver", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		testTable.Resolver = doNothingResolver
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Nil(t, err)
@@ -243,36 +251,30 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("simple returning resources insert", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
-		testTable.Resolver = dataReturningResolver
-		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		assert.Nil(t, err)
-		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
-	})
-	t.Run("simple returning resources insert w/disable_delete", func(t *testing.T) {
-		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
 		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
 		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		testTable.Resolver = dataReturningResolver
-		exec.disableDelete = true
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		exec.disableDelete = false
 		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		assert.Nil(t, err)
 	})
 	t.Run("simple returning single resources insert", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		testTable.Resolver = dataReturningSingleResolver
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Nil(t, err)
 	})
 	t.Run("simple returning nil resources insert", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		testTable.Resolver = passingNilResolver
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Nil(t, err)
@@ -280,7 +282,8 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 	})
 	t.Run("check post row resolver", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
 		testTable.Resolver = dataReturningSingleResolver
 		var expectedResource *Resource
 		testTable.PostResourceResolver = func(ctx context.Context, meta ClientMeta, parent *Resource) error {
@@ -289,7 +292,8 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 			expectedResource = parent
 			return nil
 		}
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Equal(t, expectedResource.data["name"], "other")
 		assert.Nil(t, err)
@@ -302,9 +306,11 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("test resolving with default column values", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, false, nil, false)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testDefaultsTable, nil, false)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testDefaultsTable, exec.executionStart, mock.Anything).Return(nil)
+		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			res <- testDefaultsTableData{Name: nil}
 			return nil
 		}
@@ -313,15 +319,15 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 			expectedResource = parent
 			return nil
 		}
-		_, err := execDefault.ResolveTable(context.Background(), mockedClient, nil)
+		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
 		assert.Nil(t, err)
 		assert.Equal(t, expectedResource.data["name"], "defaultValue")
 	})
 
 	t.Run("disable delete", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, true, nil, false)
-		//mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
 		testTable.Resolver = dataReturningSingleResolver
 		testTable.DeleteFilter = func(meta ClientMeta, r *Resource) []interface{} {
 			return nil
@@ -334,25 +340,27 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 			return nil
 		}
 		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
-		mockDb.On("Delete", mock.Anything, testTable, mock.Anything).Return(nil)
 		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		assert.Equal(t, expectedResource.data["name"], "other")
 		assert.Nil(t, err)
-		exec = NewExecutionData(mockDb, logger, testTable, false, nil, false)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
+
+		// new mockDb, new call counts
+		mockDb = new(DatabaseMock)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec = NewExecutionData(mockDb, logger, testTable, nil, false)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		_, err = exec.ResolveTable(context.Background(), mockedClient, nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 1)
-		mockDb.AssertNumberOfCalls(t, "CopyFrom", 2)
+		mockDb.AssertNumberOfCalls(t, "RemoveStaleData", 1)
+		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		assert.Nil(t, err)
 	})
 	t.Run("disable delete w/deleteFilter", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, true, map[string]interface{}{"test": 1}, false)
-		//mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, map[string]interface{}{"test": 1}, false)
 		testTable.Resolver = dataReturningSingleResolver
 		testTable.DeleteFilter = func(meta ClientMeta, r *Resource) []interface{} {
 			return []interface{}{"a", 2}
@@ -365,25 +373,28 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 			return nil
 		}
 		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
-		mockDb.On("Delete", mock.Anything, testTable, mock.Anything).Return(nil)
 		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, []interface{}{"test", 1, "a", 2}).Return(nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		assert.Equal(t, expectedResource.data["name"], "other")
 		assert.Nil(t, err)
-		exec = NewExecutionData(mockDb, logger, testTable, false, nil, false)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
+
+		// new mockDb, new call counts
+		mockDb = new(DatabaseMock)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec = NewExecutionData(mockDb, logger, testTable, nil, false)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		_, err = exec.ResolveTable(context.Background(), mockedClient, nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 1)
-		mockDb.AssertNumberOfCalls(t, "CopyFrom", 2)
+		mockDb.AssertNumberOfCalls(t, "RemoveStaleData", 1)
+		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		assert.Nil(t, err)
 	})
 
 	t.Run("disable delete failed copy from", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, true, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, testTable, nil, false)
 		testTable.Resolver = dataReturningSingleResolver
 		testTable.DeleteFilter = func(meta ClientMeta, r *Resource) []interface{} {
 			return nil
@@ -398,10 +409,7 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 		mockDb.On("RemoveStaleData", mock.Anything, testTable, exec.executionStart, mock.Anything).Return(nil)
 		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(fmt.Errorf("some error"))
 		mockDb.On("Insert", mock.Anything, mock.Anything, mock.Anything).Return(nil)
-		mockDb.On("Delete", mock.Anything, testTable, mock.Anything).Return(nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		_, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		mockDb.AssertNumberOfCalls(t, "Delete", 0)
 		mockDb.AssertNumberOfCalls(t, "CopyFrom", 1)
 		mockDb.AssertNumberOfCalls(t, "Insert", 1)
 		assert.Equal(t, expectedResource.data["name"], "other")
@@ -410,7 +418,8 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("always delete with disable delete", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, alwaysDeleteTable, true, nil, false)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		exec := NewExecutionData(mockDb, logger, alwaysDeleteTable, nil, false)
 		alwaysDeleteTable.Resolver = dataReturningSingleResolver
 		alwaysDeleteTable.DeleteFilter = func(meta ClientMeta, r *Resource) []interface{} {
 			return nil
@@ -432,35 +441,13 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 		assert.Nil(t, err)
 	})
 
-	t.Run("inject fields into execution", func(t *testing.T) {
-		mockDb := new(DatabaseMock)
-		exec := NewExecutionData(mockDb, logger, testTable, false, map[string]interface{}{"injected_field": 1}, false)
-		testTable.Resolver = dataReturningSingleResolver
-		testTable.DeleteFilter = nil
-		var expectedResource *Resource
-		testTable.PostResourceResolver = func(ctx context.Context, meta ClientMeta, parent *Resource) error {
-			err := parent.Set("name", "other")
-			assert.Nil(t, err)
-			expectedResource = parent
-			return nil
-		}
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, map[string]interface{}{"injected_field": 1}).Return(nil)
-		count, err := exec.ResolveTable(context.Background(), mockedClient, nil)
-		assert.Equal(t, count, uint64(1))
-		assert.Nil(t, err)
-		assert.Equal(t, expectedResource.data["name"], "other")
-		assert.Equal(t, 1, expectedResource.extraFields["injected_field"])
-		values, err := expectedResource.Values()
-		assert.Nil(t, err)
-		assert.Equal(t, []string{"name", "name_no_prefix", "prefix_name", "cq_id", "meta", "injected_field"}, expectedResource.columns)
-		assert.Equal(t, []interface{}{"other", "name_no_prefix", "prefix_name", expectedResource.cqId, expectedResource.Get("meta"), 1}, values)
-	})
-
 	t.Run("test partial fetch post resource resolver", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, false, nil, true)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testDefaultsTable, execDefault.executionStart, mock.Anything).Return(nil)
+		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			res <- testDefaultsTableData{Name: nil}
 			return nil
 		}
@@ -478,9 +465,11 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("test partial fetch resolver", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, false, nil, true)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testDefaultsTable, execDefault.executionStart, mock.Anything).Return(nil)
+		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			res <- testDefaultsTableData{Name: nil}
 			return fmt.Errorf("random failure")
 		}
@@ -498,9 +487,10 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("test partial fetch resolver panic", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, false, nil, true)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			res <- testDefaultsTableData{Name: nil}
 			panic("test panic")
 		}
@@ -518,9 +508,11 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("test partial fetch post resource resolver panic", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, false, nil, true)
-		mockDb.On("CopyFrom", mock.Anything, mock.Anything, false, mock.Anything).Return(nil)
-		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error {
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execDefault := NewExecutionData(mockDb, logger, testDefaultsTable, nil, true)
+		mockDb.On("CopyFrom", mock.Anything, mock.Anything, true, mock.Anything).Return(nil)
+		mockDb.On("RemoveStaleData", mock.Anything, testDefaultsTable, execDefault.executionStart, mock.Anything).Return(nil)
+		testDefaultsTable.Resolver = func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error {
 			res <- testDefaultsTableData{Name: nil}
 			return nil
 		}
@@ -538,7 +530,9 @@ func TestExecutionData_ResolveTable(t *testing.T) {
 
 	t.Run("test table with multiplex", func(t *testing.T) {
 		mockDb := new(DatabaseMock)
-		execDefault := NewExecutionData(mockDb, logger, testMultiplexTable, false, nil, true)
+		mockDb.On("Dialect").Return(PostgresDialect{})
+		execDefault := NewExecutionData(mockDb, logger, testMultiplexTable, nil, true)
+		mockDb.On("RemoveStaleData", mock.Anything, testMultiplexTable, execDefault.executionStart, mock.Anything).Return(nil)
 		var parentMultiplexCalled, relationMultiplexCalled = false, false
 		testMultiplexTable.Multiplex = func(meta ClientMeta) []ClientMeta {
 			parentMultiplexCalled = true