+++ provider/schema/table.go
@@ -11,7 +11,7 @@ import (
 // - parent(Resource): resource is the parent resource in case this table is called via parent table (i.e. relation)
 // - res(chan interface{}): is a channel to pass results fetched by the TableResolver
 //
-type TableResolver func(ctx context.Context, meta ClientMeta, parent *Resource, res chan interface{}) error
+type TableResolver func(ctx context.Context, meta ClientMeta, parent *Resource, res chan<- interface{}) error
 
 // IgnoreErrorFunc checks if returned error from table resolver should be ignored.
 type IgnoreErrorFunc func(err error) bool
@@ -24,7 +24,7 @@ type Table struct {
 	// table description
 	Description string
 	// Columns are the set of fields that are part of this table
-	Columns []Column
+	Columns ColumnList
 	// Relations are a set of related tables defines
 	Relations []*Table
 	// Resolver is the main entry point to fetching table data and
@@ -42,18 +42,10 @@ type Table struct {
 	// AlwaysDelete will always delete table data on fetch regardless if delete is disabled on run,
 	// use this only in specific cases, if you are unsure contact the CloudQuery Team.
 	AlwaysDelete bool
-}
-
-// ColumnNames returns all collected columns name of table (including all inner embedded columns)
-func (t Table) ColumnNames() []string {
-	var cn = make([]string, len(t.Columns))
-	for i, c := range t.Columns {
-		cn[i] = c.Name
-	}
-	for _, c := range GetDefaultSDKColumns() {
-		cn = append(cn, c.Name)
-	}
-	return cn
+	// IgnoreInTests if true this skips this column in tests as sometimes it might be hard
+	// to create a reproducible test environment with this column being non nil. For example various error tables such as
+	// security violations and so.
+	IgnoreInTests bool
 }
 
 func (t Table) Column(name string) *Column {
@@ -65,16 +57,8 @@ func (t Table) Column(name string) *Column {
 	return nil
 }
 
-func (t Table) PrimaryKeys() []string {
-	if len(t.Options.PrimaryKeys) > 0 {
-		return t.Options.PrimaryKeys
-	}
-	return []string{"cq_id"}
-}
-
 // TableCreationOptions allow modifying how table is created such as defining primary keys, indices, foreign keys and constraints.
 type TableCreationOptions struct {
-	// List of columns to set as primary keys, if HashPrimaryKeys is true the column values will be used to generate an ID
-	// and an "id" column will be created for the table. If this nil a random unique ID is generated.
+	// List of columns to set as primary keys. If this is empty, a random unique ID is generated.
 	PrimaryKeys []string
 }